Add support for running 2019 Worst Evictors sql code  (#3)
This PR adds a command for re-running our sql code to generate the raw data for the 2019 Worst Evictors list. It also renames a bunch of files and commands to make the distinction between 2018 and 2019 more clear. Hopefully, too, this will allow us to add handling for future years more easily!